+++ gamemodes/terrortown/gamemode/client/cl_changes.lua
@@ -2304,6 +2304,62 @@ function CreateChanges()
         os.time({ year = 2025, month = 02, day = 02 })
     )
 
+    AddChange(
+        "TTT2 Base - v0.14.3b",
+        [[
+
+        <h2>Added</h2>
+        <ul>
+            <li>Added C4 cvars for adjusting the range & falloff of C4 explosions (by @Spanospy)</li>
+            <ul>
+                <li>C4 explosions will now also respect the weapon's configured damage scaling</li>
+            </ul>
+            <li>Added option to select preferred unit of length for distance displays (by @wgetJane)</li>
+            <li>Added <code>GM:TTT2ArmorHandlePlayerTakeDamage</code> hook for modifying/overriding armor behavior (by @wgetJane)</li>
+            <li>Added server option for body armor to protect against crowbar damage (by @wgetJane)</li>
+            <li>Added <code>GM:TTTLastWordsMsg</code> hook from base TTT (by @wgetJane)</li>
+            <li>Port new TTT entity <code>ttt_filter_role</code> to TTT2 (by @figardo, ported by @wgetJane)</li>
+            <li>Added 2 new sprint settings (by @wgetJane)</li>
+            <ul>
+                <li>Stamina cooldown time before stamina begins regenerating after sprinting, 0.8 seconds by default</li>
+                <li>Option to require forward key to be held to sprint, enabled by default</li>
+            </ul>
+        </ul>
+
+        <h2>Changed</h2>
+        <ul>
+            <li>Updated Russian and English localization files (by @Satton2)</li>
+            <li>Updated the list of troublesome addons used by the addonchecker</li>
+            <li>Changed option for body armor to protect against headshot damage by default (by @wgetJane)</li>
+        </ul>
+
+        <h2>Fixed</h2>
+        <ul>
+            <li>Fixed classic armour protecting against crowbar damage (by @wgetJane)</li>
+            <li>Fixed C4/Radio sounds not playing outside of PAS (by @figardo)</li>
+            <li>Fixed players sometimes being revealed as dead when they chat/voicechat right as they die (by @wgetJane)</li>
+            <li>Fixed various bugs related to using doors and buttons, to match base TTT behavior (by @wgetJane)</li>
+            <ul>
+                <li>Fixed brush doors on certain maps not being usable</li>
+                <li>Fixed certain doors on certain maps being forcibly usable when they shouldn't be</li>
+                <li>Fixed func_rotating being forcibly usable</li>
+                <li>Fixed vehicles not being usable</li>
+                <li>Removed forced button solidity</li>
+                <li>Improved targetID trace detection for doors and buttons</li>
+                <li>Allow brush doors to be detected as destructible for targetID</li>
+                <li>Use NW2Vars instead of NWVars for door variables for more efficient networking</li>
+                <li>Fixed clientside door.GetAll() to return a more updated list of doors</li>
+                <li>Removed clientside TTT2PostDoorSetup hook</li>
+                <li>Changed crowbar unlocking behavior to match base TTT</li>
+            </ul>
+            <li>Fixed lua errors after autorefresh (by @wgetJane)</li>
+            <li>Fixed Hud errors when picking up items or weapons with no viable icon (by @NickCloudAT)</li>
+        </ul>
+
+    ]],
+        os.time({ year = 2025, month = 03, day = 18 })
+    )
+
     ---
     -- run hook for other addons to add their changelog as well
     -- @realm client